New SimpleNestedIF

+++ SimpleIfElseIf.kt
@@ -1,10 +1,10 @@
 fun main(args:Array<String>){
     print("Nilai : ")
-    var nilai:Double= readLine()!!.toDouble()
+    var nilai:Int= readLine()!!.toInt()
 
     if (nilai<75){
         print("Anda Harus Remedial!")
-    }else if(nilai==75){
+    }else if(nilai == 75){
         println("Selamat Anda Lulus!")
         print("Nilai anda pas dengan KKM!")
     }else{
